# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Todoapp local server",
        "title": "Todoapp API",
        "contact": {}
    },
    "host": "localhost:3000",
    "paths": {
        "/": {
            "get": {
                "description": "retrieves a list of all tasks in database in json format",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "gets all tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TodoItem"
                            }
                        }
                    },
                    "400": {
                        "description": "error retrieving all tasks"
                    }
                }
            }
        },
        "/todo/": {
            "post": {
                "description": "accepts task's title and completed bool in json format and returns the created task in json",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "inserts new task",
                "parameters": [
                    {
                        "description": "task body",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TodoItem"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.TodoItem"
                        }
                    },
                    "400": {
                        "description": "error binding json data"
                    },
                    "500": {
                        "description": "error inserting new task"
                    }
                }
            }
        },
        "/todo/:id": {
            "get": {
                "description": "retrieves a list of all tasks in database in json format",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "gets a task by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "required task's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TodoItem"
                        }
                    },
                    "400": {
                        "description": "error retrieving task"
                    }
                }
            },
            "delete": {
                "description": "accepts task's id and deletes it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "delete"
                ],
                "summary": "deletes a task",
                "parameters": [
                    {
                        "description": "required task's id",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TodoItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error deleting task"
                    }
                }
            },
            "patch": {
                "description": "accepts task's title and completed bool in json format and returns the created task in json",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "updates a task",
                "parameters": [
                    {
                        "description": "required task's id",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TodoItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error binding json data"
                    },
                    "500": {
                        "description": "error updating task"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.TodoItem": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "task": {
                    "type": "string"
                }
            }
        }
    }
}